Create separate components to overview page and make use of useContext hook in each component
